+++ sdk/identity/azure-identity/src/environment_credential.cpp
@@ -5,6 +5,7 @@
 
 #include "azure/identity/client_certificate_credential.hpp"
 #include "azure/identity/client_secret_credential.hpp"
+#include "azure/identity/detail/client_credential_core.hpp"
 #include "private/identity_log.hpp"
 
 #include <azure/core/azure_assert.hpp>
@@ -28,7 +29,6 @@ namespace {
 constexpr auto AzureTenantIdEnvVarName = "AZURE_TENANT_ID";
 constexpr auto AzureClientIdEnvVarName = "AZURE_CLIENT_ID";
 constexpr auto AzureClientSecretEnvVarName = "AZURE_CLIENT_SECRET";
-constexpr auto AzureAuthorityHostEnvVarName = "AZURE_AUTHORITY_HOST";
 constexpr auto AzureClientCertificatePathEnvVarName = "AZURE_CLIENT_CERTIFICATE_PATH";
 
 void PrintCredentialCreationLogMessage(
@@ -46,7 +46,7 @@ EnvironmentCredential::EnvironmentCredential(
   auto clientId = Environment::GetVariable(AzureClientIdEnvVarName);
 
   auto clientSecret = Environment::GetVariable(AzureClientSecretEnvVarName);
-  auto authority = Environment::GetVariable(AzureAuthorityHostEnvVarName);
+  auto authority = Environment::GetVariable(_detail::AzureAuthorityHostEnvVarName);
 
   auto clientCertificatePath = Environment::GetVariable(AzureClientCertificatePathEnvVarName);
 
@@ -65,7 +65,7 @@ EnvironmentCredential::EnvironmentCredential(
 
       if (!authority.empty())
       {
-        envVarsToParams.push_back({AzureAuthorityHostEnvVarName, "authorityHost"});
+        envVarsToParams.push_back({_detail::AzureAuthorityHostEnvVarName, "authorityHost"});
         clientSecretCredentialOptions.AuthorityHost = authority;
       }
 
@@ -85,7 +85,7 @@ EnvironmentCredential::EnvironmentCredential(
 
       if (!authority.empty())
       {
-        envVarsToParams.push_back({AzureAuthorityHostEnvVarName, "authorityHost"});
+        envVarsToParams.push_back({_detail::AzureAuthorityHostEnvVarName, "authorityHost"});
         clientCertificateCredentialOptions.AuthorityHost = authority;
       }
 
@@ -109,7 +109,7 @@ EnvironmentCredential::EnvironmentCredential(
       auto logMsg = GetCredentialName() + ": Both '" + AzureTenantIdEnvVarName + "' and '"
           + AzureClientIdEnvVarName + "', and at least one of '" + AzureClientSecretEnvVarName
           + "', '" + AzureClientCertificatePathEnvVarName + "' needs to be set. Additionally, '"
-          + AzureAuthorityHostEnvVarName
+          + _detail::AzureAuthorityHostEnvVarName
           + "' could be set to override the default authority host. Currently:\n";
 
       std::pair<char const*, bool> envVarStatus[] = {
@@ -117,7 +117,7 @@ EnvironmentCredential::EnvironmentCredential(
           {AzureClientIdEnvVarName, !clientId.empty()},
           {AzureClientSecretEnvVarName, !clientSecret.empty()},
           {AzureClientCertificatePathEnvVarName, !clientCertificatePath.empty()},
-          {AzureAuthorityHostEnvVarName, !authority.empty()},
+          {_detail::AzureAuthorityHostEnvVarName, !authority.empty()},
       };
       for (auto const& status : envVarStatus)
       {