+++ sdk/identity/azure-identity/CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Bugs Fixed
 
+- Change the default value for the authority host option to be read from the environment variable first.
+
 ### Other Changes
 
 ## 1.6.0-beta.2 (2023-09-13)